[Feature Branch] Resolve review comments (#1154)
